improve image handling
